pyproject.toml
[project]
name = "mc-llm-utils"
version = "0.9.5"
description = "Lightweight utilities for LLM data parsing and training"
authors = [
    { name="Mitchell Currie" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx"
]

[project.optional-dependencies]
training = [
    "transformers>=4.41.0",
    "scikit-learn",
    "pandas",
    "torch",
    "datasets",
    "numpy",
    "tqdm",
    "sentence-transformers",
    "pynvml",
    "urllib3",
    "safetensors",
    "tensorboard",
    "evaluate",
    "psutil",
    "accelerate"
]
test = [
    "pytest",
    "numpy",
    "httpx"
]
google = [
    "google-generativeai",
]

[project.urls]
"Homepage" = "https://github.com/mitchins/llm-utils"


[tool.setuptools.packages.find]
include = ["llm_utils*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
train-bert = "llm_utils.training.train_bert:main"
run-bert = "llm_utils.inference.run_bert:main"
train-t5 = "llm_utils.training.train_t5:main"
run-t5 = "llm_utils.inference.run_t5:main"
prepare-dataset = "llm_utils.training.prepare_dataset:main"
llm-request = "llm_utils.clients.openai_client:main"
